// attached_assets/upgrade/types.ts
export type ToastType = 'success' | 'error' | 'info';

export interface ToastMessage {
  id: number;
  message: string;
  type: ToastType;
  onClick?: () => void;
}